src/streams/src/main/python/requirements.txt
sseclient==0.0.27
kafka-python==2.0.2
dataclasses-jsonschema==2.15.1
